add dao, store

+++ app/src/main/java/com/example/currencyexchange/models/data/Money.kt
@@ -1,6 +1,6 @@
 package com.example.currencyexchange.models.data
 
 data class Money(
-    val amount: Int,
+    val amount: Float,
     val currencyCode: String
 )
